support skopeo

+++ syncer.py
@@ -10,17 +10,18 @@ helpInfo = 'syncer.py -s https://registry1.xxx.xx -t registry2.xxx.xx:6060'
 def parseUrl(uri):
     matchObj = re.match( r'(https?://)(.*)', uri, re.M|re.I)
     if matchObj:
-      print matchObj.group(1),'----',matchObj.group(2)
       return matchObj.group(1),matchObj.group(2)
     else:
-      print uri+' must start with http or https'
+      print(uri+' must start with http or https')
       sys.exit()
-def main(argv):
+
+def parseOpts(argv):
   sProtocol=''
   source = ''
   target = ''
+  model = ''
   try:
-    opts, args = getopt.getopt(argv,"hs:t:",["source=","target="])
+    opts, _ = getopt.getopt(argv,"hs:t:",["source=","target="])
   except getopt.GetoptError:
     print(helpInfo)
     sys.exit(2)
@@ -32,9 +33,15 @@ def main(argv):
        (sProtocol,source) = parseUrl(arg)
     elif opt in ("-t", "--target"):
        target = arg
-  print source,target
+    elif opt in ("-m", "--model"):
+       model= arg
+  print (source,target)
+  return sProtocol,source,target,model
+
+def main(argv):
+  (sProtocol,source,target,model) = parseOpts(argv)
   val = commands.getoutput('curl -k -s '+sProtocol+source+'/v2/_catalog')
-  print val
+  print(val)
   load_json=json.loads(val)
   for i in load_json["repositories"]:
       (status,val)=commands.getstatusoutput('curl -s -k '+sProtocol+source+'/v2/'+i+'/tags/list')
@@ -48,8 +55,12 @@ def main(argv):
             if len(tokens) == 1:
               imageUri='library/'+imageUri
             targetUri=target+'/'+imageUri
-            cmd_str='docker pull '+sourceUri+' && docker tag '+sourceUri+' '+targetUri
-            cmd_str=cmd_str+' && docker push '+targetUri+' && docker rmi '+sourceUri+' && docker rmi '+targetUri
+            cmd_str = ''
+            if model == 'sk':
+              cmd_str='skopeo copy docker://'+sourceUri+' docker://'+targetUri
+            else:
+              cmd_str='docker pull '+sourceUri+' && docker tag '+sourceUri+' '+targetUri
+              cmd_str=cmd_str+' && docker push '+targetUri+' && docker rmi '+sourceUri+' && docker rmi '+targetUri
             print(cmd_str)
             os.system(cmd_str)
